requirements.txt
langgraph>=0.0.15
google-generativeai>=0.3.1
langchain>=0.1.0
langchain-google-genai>=0.0.5
typing-extensions>=4.8.0 